Корректировка под критери
