+++ package.json
@@ -18,6 +18,7 @@
     "gulp-sass": "4.0.x",
     "gulp-sourcemaps": "2.6.x",
     "gulp-svgstore": "^7.0.1",
+    "gulp-uglify": "^3.0.2",
     "gulp-webp": "^4.0.1",
     "install": "^0.12.2",
     "posthtml-include": "^1.2.0",
@@ -48,5 +49,7 @@
   "engines": {
     "node": "8.15"
   },
-  "dependencies": {}
+  "dependencies": {
+    "polyfill": "^0.1.0"
+  }
 }
