+++ source/sass/catalog-button.scss
@@ -27,7 +27,6 @@
   }
 }
 
-
 @media (min-width: $tablet-width) {
   .catalog-button {
     padding: 13px 91px 12px 95px;
