+++ source/sass/contacts.scss
@@ -15,7 +15,7 @@
 .contacts__wrapper {
   display: flex;
   width: 276px;
-  padding: 33px 0 28px;
+  padding: 22px 0 28px;
   text-align: left;
   margin: 0 auto;
   z-index: 3;
@@ -63,9 +63,8 @@
   font-family: $font-arial;
   font-size: 14px;
   line-height: 1.43;
-  text-align: right;
+  text-align: left;
 }
-
 
 @media (min-width: $tablet-width) {
   .contacts {
@@ -76,7 +75,7 @@
 
   .contacts__wrapper {
     width: 768px;
-    padding: 69px 64px 59px 41px;
+    padding: 59px 64px 59px 64px;
     box-sizing: border-box;
   }
