+++ source/sass/description.scss
@@ -1,10 +1,11 @@
 .description {
   margin: 0 auto;
-  margin-bottom: 43px;
+  margin-bottom: 9px;
 }
 
 .description__wrapper {
   margin: 0 auto;
+  width: 320px;
 }
 
 .description__title {
@@ -15,7 +16,8 @@
   color: $second-color;
   font-weight: 500;
   padding-left: 20px;
-  margin-bottom: 28px;
+  margin-bottom: 24px;
+  word-break: break-all;
 }
 
 .description__list {
@@ -144,8 +146,8 @@
     }
 
     &--advantage,
-    &--enjoy {
-      margin-bottom: 123px;
+    &--powder {
+      margin-bottom: 132px;
     }
 
     &--advantage,
@@ -155,6 +157,10 @@
 
     &--powder {
       width: 230px;
+    }
+
+    &--enjoy {
+      margin-bottom: 104px;
     }
   }
 }
