+++ source/sass/logo-link.scss
@@ -26,6 +26,7 @@
   .logo-link {
     width: 128px;
     margin: 0;
+    margin-right: 150px;
     border-bottom: none;
     padding-top: 24px;
   }
