+++ source/sass/page-header.scss
@@ -5,15 +5,13 @@
   margin: 0 auto;
 }
 
-.page-header__wrapper {
-  padding-top: 13px;
-}
-
 .page-header__logo {
   width: 191px;
-  height: 54px;
+  height: 60px;
   padding-left: 20px;
   cursor: pointer;
+  display: inline-block;
+  padding-top: 10px;
 }
 
 .page-header a[href] {
@@ -35,8 +33,8 @@
 
 @media (min-width: $tablet-width) {
   .page-header {
-    width: 708px;
-    margin: 22px auto 56px;
+    width: 750px;
+    margin: 22px auto 36px;
     flex-direction: row;
   }
 
@@ -61,7 +59,7 @@
 @media (min-width: $desktop-width) {
   .page-header {
     width: 1300px;
-    margin: 0 auto 71px;
+    margin: 0 auto 59px;
     box-sizing: border-box;
     position: relative;
 
@@ -92,7 +90,7 @@
 
   .page-header__wrapper {
     width: 1300px;
-    padding-left: 69px;
+    padding-left: 38px;
     padding-top: 53px;
     margin: 0 auto;
     box-sizing: border-box;
