+++ source/sass/page-main.scss
@@ -27,7 +27,7 @@
     position: absolute;
     width: 552px;
     height: 499px;
-    top: 152px;
-    left: 39%;
+    top: 140px;
+    left: 38%;
   }
 }
